backend/requirements.txt
aiohttp==3.10.11
aiohttp-security==0.5.0
PyJWT==2.10.1
